+++ tests/algorithms/multiway/test_matchals.py
@@ -1,0 +1,45 @@
+from unittest import TestCase
+
+import numpy as np
+import networkx as nx
+
+import graph_matching_tools.algorithms.multiway.matchals as matchals
+import graph_matching_tools.algorithms.kernels.gaussian as kern
+import graph_matching_tools.algorithms.kernels.utils as utils
+
+
+class TestMatchALS(TestCase):
+    def test_matchals(self):
+        node_kernel = kern.create_gaussian_node_kernel(2.0, "weight")
+
+        graph1 = nx.Graph()
+        graph1.add_node(0, weight=2.0)
+        graph1.add_node(1, weight=20.0)
+        graph1.add_edge(0, 1, weight=1.0)
+
+        graph2 = nx.Graph()
+        graph2.add_node(0, weight=20.0)
+        graph2.add_node(1, weight=2.0)
+        graph2.add_edge(0, 1, weight=1.0)
+
+        graph3 = nx.Graph()
+        graph3.add_node(0, weight=-5.0)
+        graph3.add_node(1, weight=30.0)
+        graph3.add_node(2, weight=2.0)
+        graph3.add_edge(1, 2, weight=1.0)
+        graphs = [graph1, graph2, graph3]
+
+        knode = utils.create_full_node_affinity_matrix(graphs, node_kernel)
+        res = matchals.matchals(knode, [2, 2, 3], 4, alpha=200.0, random_seed=11)
+
+        truth = [
+            [1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
+            [0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0],
+            [0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0],
+            [1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
+            [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0],
+            [0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0],
+            [1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
+        ]
+
+        self.assertEqual(np.linalg.norm(res - truth) < 1e-3, True)